refactor: добавил в логику формирования фильма режиссеров. feature: реализовал сортировку популярных фильмов по режиссеру в findFilmsByDirectorSortedByYear

+++ src/main/java/ru/yandex/practicum/filmorete/service/ServiceFilm.java
@@ -81,8 +81,8 @@ public class ServiceFilm {
             }
         }
 
-        if (film.getDirector() != null) {
-            for (Director director : film.getDirector()) {
+        if (film.getDirectors() != null) {
+            for (Director director : film.getDirectors()) {
                 totalDirectorFilmDao.insert(optionalFilm.get().getId(), director.getId());
             }
         }
@@ -99,9 +99,10 @@ public class ServiceFilm {
                     film.getReleaseDate(), film.getDuration()
             );
 
-            if (film.getDirector() != null) {
-                totalDirectorFilmDao.deleteAllByFilmId(film.getId());
-                for (Director director : film.getDirector()) {
+            List<TotalDirectorFilm> totalDirectorFilms = totalDirectorFilmDao.findAllTotalDirectorFilm(film.getId());
+            if (!totalDirectorFilms.isEmpty()) totalDirectorFilmDao.deleteAllByFilmId(film.getId());
+            if (film.getDirectors() != null) {
+                for (Director director : film.getDirectors()) {
                     totalDirectorFilmDao.insert(film.getId(), director.getId());
                 }
             }
@@ -140,6 +141,14 @@ public class ServiceFilm {
         if (optionalFilm.isEmpty()) throw new ExceptionNotFoundFilmStorage(VALID_ERROR_FILM_ID_NOT_IN_COLLECTIONS);
         if (optionalUser.isEmpty()) throw new ExceptionNotFoundUserStorage(VALID_ERROR_USER_ID_NOT_IN_COLLECTIONS);
         totalFilmLikeDao.insert(filmId, userId);
+    }
+
+    public List<Film> getFilmsToDirector(Long directorId, String sorted) {
+        if (sorted.equals("year")) {
+            return totalDirectorFilmDao.findFilmsByDirectorSortedByYear(directorId);
+        } else {
+            return totalDirectorFilmDao.findPopularFilmsByDirector(directorId);
+        }
     }
 
     public void clearStorage() {
